io_ddir: fix -Wsign-compare in io_ddir_name()
ddir is a signed enum fio_ddir while sizeof(name)/sizeof(name[0]) is
unsigned long, which trips -Werror=sign-compare in downstream builds
(e.g. the SPDK fio_plugin). Cast the array size to int; the ddir >= 0
guard still protects the negative enum values.

+++ io_ddir.h
@@ -33,7 +33,7 @@ static inline const char *io_ddir_name(enum fio_ddir ddir)
 		[DDIR_WAIT] = "wait",
 	};
 
-	if (ddir >= 0 && ddir < sizeof(name) / sizeof(name[0]) && name[ddir])
+	if (ddir >= 0 && ddir < (int)(sizeof(name) / sizeof(name[0])) && name[ddir])
 		return name[ddir];
 
 	return "invalid";
